more as-provider persistence

+++ server/migrations/20240709141015_create_qs_queue_data_table.sql
@@ -7,6 +7,6 @@
 CREATE TABLE qs_queue_data(
 queue_id uuid NOT NULL,
 PRIMARY KEY (queue_id),
-sequence_number NUMERIC NOT NULL,
+sequence_number BIGINT NOT NULL,
 FOREIGN KEY (queue_id) REFERENCES qs_client_records(client_id) ON DELETE CASCADE
 );
